ETCH-132 Major changes in platform abstraction
Restructuring of os specific file hierarchy
Added QNX, Linux ARMv7 and Integrity ARMv7 support

Added Array, Ringbuffer and Queue containers
Added Math abstraction

Fixed warnings on windows Visual Studio build


git-svn-id: https://svn.apache.org/repos/asf/incubator/etch/trunk@1368491 13f79535-47bb-0310-9956-ffa450edef68

+++ binding-cpp/runtime/include/common/EtchDate.h
@@ -59,7 +59,7 @@ public:
   /**
    * Returns hash code
    */
-  capu::uint64_t getHashCode() const;
+  capu::uint32_t getHashCode() const;
 
 private:
