+++ binding-cpp/runtime/include/common/EtchObjectHash.h
@@ -24,15 +24,15 @@
 class EtchObjectHash {
 public:
 
-  static capu::uint64_t Digest(const EtchObject &key) {
+  static capu::uint32_t Digest(const EtchObject &key) {
     return key.getHashCode();
   }
 
-  static capu::uint64_t Digest(const EtchObject* key) {
+  static capu::uint32_t Digest(const EtchObject* key) {
     return key->getHashCode();
   }
 
-  static capu::uint64_t Digest(const capu::SmartPointer<EtchObject>& key) {
+  static capu::uint32_t Digest(const capu::SmartPointer<EtchObject>& key) {
     return key->getHashCode();
   }
 };
